Bug 1849599 - Reorganize the composed Smoke tests suite
